+++ cafana/include/pi02024/utilities_pi02024_phase.h
@@ -55,7 +55,7 @@ struct reco_inter_phase {
   double pi0_leading_photon_cosphi;
   double pi0_leading_photon_conv_dist;
   double pi0_subleading_photon_energy;
-  size_t pi0_subleading_photon_cosphi;
+  double pi0_subleading_photon_cosphi;
   double pi0_subleading_photon_conv_dist;
   double pi0_costheta;
   double pi0_mass;
